Added compressed SVGs and relocated to static folder, fixed styling for Teams Card and added line spacing and formatting
