+++ waterloop-site/src/components/SustainableTech/Teams/TeamsCarousel.tsx
@@ -1,4 +1,5 @@
 import React from 'react';
+
 import { useSwipeable } from 'react-swipeable';
 import styled from 'styled-components';
 import useTeamImages from './hooks/team-images';
@@ -16,22 +17,25 @@ const TeamCardContainer = styled.div`
   display: flex;
   justify-content: center;
   align-items: center;
+  max-height: 380px;
+  min-height: 380px;
   margin: 0 30px;
   padding: 40px 20px;
-  background-color: #618A4D;
+  background-color: #618a4d;
   box-shadow: 0px 4px 40px rgba(0, 0, 0, 0.15);
-  border-radius: 8px;
+  border-radius: 20px;
   @media screen and (max-width: 768px) {
     gap: 10px;
     flex-direction: column;
     padding: 40px;
+    max-height: none;
   }
 `;
 
 const SlideContainer = styled.div`
   display: flex;
   width: 100%;
-  gap: 30px;
+  gap: 50px;
   justify-content: center;
   align-items: center;
   @media screen and (max-width: 768px) {
@@ -41,7 +45,10 @@ const SlideContainer = styled.div`
 `;
 
 const TeamImageWrapper = styled.div`
-  max-width: 300px;
+  max-width: 40%;
+  @media screen and (max-width: 768px) {
+    max-width: 90%;
+  }
 `;
 
 const TeamImage = styled.img`
@@ -49,8 +56,16 @@ const TeamImage = styled.img`
   border-radius: 8px;
 `;
 
+const TeamInfoWrapper = styled.div`
+  max-width: 55%;
+  @media screen and (max-width: 768px) {
+    max-width: 90%;
+  }
+`;
+
 const TeamInfoHeading = styled.h2`
   font-weight: bold;
+  font-style: normal;
   font-size: 20px;
   color: #ffffff;
 `;
@@ -110,7 +125,7 @@ const TeamsPickerImage = styled.img`
 const TeamsPickerCard = styled.div`
   /* width is set to 19% to space 5 images equally
   inside a box of width = 1200px and gap = 15px */
-  width: 19%; 
+  width: 19%;
   max-width: 150px;
   min-width: 100px;
 
@@ -120,7 +135,15 @@ const TeamsPickerCard = styled.div`
 `;
 
 const TeamsCarousel: React.FC = () => {
-  const { image, name, desc, imgs, currentTeam, cycleLeft, cycleRight, selectTeam } = useTeamImages();
+  const {
+    image,
+    name,
+    desc,
+    imgs,
+    cycleLeft,
+    cycleRight,
+    selectTeam,
+  } = useTeamImages();
 
   const swipeHandlers = useSwipeable({
     onSwipedLeft: cycleLeft,
@@ -139,12 +162,12 @@ const TeamsCarousel: React.FC = () => {
             <TeamImage src={image}></TeamImage>
           </TeamImageWrapper>
 
-          <div>
+          <TeamInfoWrapper>
             <TeamInfoHeading>{name}</TeamInfoHeading>
             <TeamInfoText>{desc}</TeamInfoText>
-          </div>
+          </TeamInfoWrapper>
         </SlideContainer>
-      
+
         <Arrow className="material-icons" onClick={cycleRight}>
           keyboard_arrow_right
         </Arrow>
@@ -161,11 +184,11 @@ const TeamsCarousel: React.FC = () => {
 
       <TeamsPickerContainer>
         {imgs.map((team, i: number) => {
-            return (
-              <TeamsPickerCard onClick={selectTeam(i)}>
-                <TeamsPickerImage src={team.imgFile} />
-              </TeamsPickerCard>
-            );
+          return (
+            <TeamsPickerCard onClick={selectTeam(i)}>
+              <TeamsPickerImage src={team.imgFile} />
+            </TeamsPickerCard>
+          );
         })}
       </TeamsPickerContainer>
     </TeamsContentWrapper>
