+++ waterloop-site/src/pages/SustainableTechEvent.tsx
@@ -1,20 +1,20 @@
 import React from 'react';
 
 import { AboutEvent } from '../components/SustainableTech/AboutEvent';
-import { Teams } from '../components/SustainableTech/Teams';
 import { Contact } from '../components/SustainableTech/Contact';
 import { FAQ } from '../components/SustainableTech/FAQ';
 import { Hero } from '../components/SustainableTech/Hero';
 import { Schedule } from '../components/SustainableTech/Schedule';
 import { Sponsors } from '../components/SustainableTech/Sponsors';
+import { Teams } from '../components/SustainableTech/Teams';
 
 const SustainableTechEvent: React.FC = () => (
   <div>
     <Hero />
     <AboutEvent />
     <Schedule />
-    <Sponsors />
     <Teams />
+    <Sponsors />
     <FAQ />
     <Contact />
   </div>
